requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
greenlet==1.1.3.post0
mysqlclient==2.1.1
SQLAlchemy==1.4.42
sqlparse==0.4.3
